created logRequest function in logger middleware
